move some bids components to bids folder
