+++ src/components/dashboard/containers/Bids/BidsCommon.js
@@ -15,7 +15,7 @@ import TableCell from '@material-ui/core/TableCell'
 import WithDialog from 'components/common/dialog/WithDialog'
 import classnames from 'classnames'
 import { withReactRouterLink } from 'components/common/rr_hoc/RRHoc'
-import ItemIpfsDetails from './ItemIpfsDetails'
+import ItemIpfsDetails from './../ItemIpfsDetails'
 import { Button, IconButton } from 'react-toolbox/lib/button'
 import { getBidEvents } from 'services/adex-node/actions'
 import { AcceptBid, GiveupBid, VerifyBid, CancelBid, RefundBid } from 'components/dashboard/forms/web3/transactions'
