+++ src/components/dashboard/containers/Bids/SlotBids.js
@@ -6,7 +6,7 @@ import { bindActionCreators } from 'redux'
 import actions from 'actions'
 import theme from './theme.css'
 import RTButtonTheme from 'styles/RTButton.css'
-import ListWithControls from './Lists/ListWithControls'
+import ListWithControls from 'components/dashboard/containers/Lists/ListWithControls'
 import Rows from 'components/dashboard/collection/Rows'
 import { Tab, Tabs } from 'react-toolbox'
 import { Grid, Row, Col } from 'react-flexbox-grid'
