+++ src/components/dashboard/containers/Bids/UnitBids.js
@@ -5,7 +5,7 @@ import { bindActionCreators } from 'redux'
 import actions from 'actions'
 import theme from './theme.css'
 import { Tab, Tabs } from 'react-toolbox'
-import ListWithControls from './Lists/ListWithControls'
+import ListWithControls from 'components/dashboard/containers/Lists/ListWithControls'
 import Rows from 'components/dashboard/collection/Rows'
 import Translate from 'components/translate/Translate'
 import classnames from 'classnames'
