+++ src/components/dashboard/containers/Slot.js
@@ -6,7 +6,7 @@ import actions from 'actions'
 import { AdSlot } from 'adex-models'
 import ItemHoc from './ItemHoc'
 import theme from './theme.css'
-import SlotBids from './SlotBids'
+import SlotBids from './Bids/SlotBids'
 import { items as ItemsConstants } from 'adex-constants'
 import { BasicProps } from './ItemCommon'
 import Helper from 'helpers/miscHelpers'
