+++ src/components/dashboard/containers/Unit.js
@@ -11,7 +11,7 @@ import UnitTargets from './UnitTargets'
 import Translate from 'components/translate/Translate'
 import WithDialog from 'components/common/dialog/WithDialog'
 import { NewBidSteps } from 'components/dashboard/forms/bids/bids'
-import UnitBids from './UnitBids'
+import UnitBids from './Bids/UnitBids'
 import { items as ItemsConstants } from 'adex-constants'
 import { BasicProps } from './ItemCommon'
 import { getUnitBids } from 'services/adex-node/actions'
